Revert "Don't sync the stored items for Handy Bag, Inventory Swapper or Memory Cards"
This reverts commit 7fe928e32f8501b23d0f52636b369588b8918041.

+++ src/main/java/fi/dy/masa/enderutilities/util/EUStringUtils.java
@@ -63,7 +63,7 @@ public class EUStringUtils
      * @param value
      * @return
      */
-    public static String formatNumberWithKSeparators(long value)
+    public static String formatNumberWithKSeparators(int value)
     {
         return NumberFormat.getNumberInstance(Locale.US).format(value);
     }
